fix: create spell on actor when dragging to embedded item
When dragging a spell onto an embedded item's spells tab, the spell
was only added to the item's flag configuration but not actually
created on the actor. This meant the spell wouldn't show up in the
actor's spell list.

Now when a spell is dropped on an embedded item, it immediately
creates the spell on the actor and updates the flag to reference
the created spell instead of the compendium spell.